% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{addFeedbackLoops}
\alias{addFeedbackLoops}
\title{Add feedback loops from target genes to proteins/TFs if present.}
\usage{
addFeedbackLoops(ST_net_targets)
}
\arguments{
\item{ST_net_targets}{full consensus graph.}
}
\value{
igraph object with feedback loops added.
}
\description{
Add feedback loops from target genes to proteins/TFs if present.
}
\keyword{manip}
