% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{consDynamicNet}
\alias{consDynamicNet}
\title{Dynamic analysis.}
\usage{
consDynamicNet(data_omics, consensusGraphs, laghankel = 3,
  cutoffhankel = 0.9, conv.1 = 0.15, conv.2 = 0.05, conv.3 = 0.05,
  verbose = TRUE, max.iter = 100, max.subiter = 200)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{consensusGraphs}{result from static analysis: consensus graph generated
by staticConsensusNet function.}

\item{laghankel}{integer specifying the maximum relevant time lag to be used
in constructing the block-Hankel matrix.}

\item{cutoffhankel}{cutoff to determine desired percent of total variance
explained; default = 0.9 as in [1].}

\item{conv.1}{value of convergence criterion 1; default value is 0.15
(for further details see [1]).}

\item{conv.2}{value of convergence criterion 2; default value is 0.05
(for further details see [1]).}

\item{conv.3}{value of convergence criterion 3; default value is 0.05
(for further details see [1]).}

\item{verbose}{boolean value, verbose output TRUE or FALSE}

\item{max.iter}{maximum overall iterations; default value is 100
(for further details see [1]).}

\item{max.subiter}{maximum iterations for hyperparameter updates;
default value is 200 (for further details see [1]).}
}
\value{
list of 2 elements:
1) output parameters of dynamic network inference with ebdbNet package
2) splines data generated.
}
\description{
Generates continous data for dynamic analysis of protein, TF and gene data
via smoothing splines. 50 time points are generated this way.
The following nodes are considered:
Nodes which are part of the static consensus graphs from corresponding time
points of the two different measurement types. In case a node is not
significantly changed at a certain point in time its FC is assumed to remain
constant at this time point.
Calculation of the consensus-based dynamic net parameters are based on the ebdbNet
R package [1]. The number of time points generated via smoothing splines (50)
is based on their results for median AUCs of ROC curves.
The number of forward time units a node is assumed to influence other nodes
can be specified via the laghankel parameter. The cutoff determining the
percent of total variance explained by the singular values generated by
singular value decomposition (SVD) of the block-Hankel matrix H in order to
specify the hidden state dimension K (for further details see [1]).
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE,
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
statConsNet = staticConsensusNet(data_omics)
consDynamicNet(data_omics, statConsNet)
}
}
\references{
1. A. Rau, F. Jaffrezic, J.-L. Foulley, R. W. Doerge (2010).
An empirical Bayesian method for estimating biological networks from temporal
microarray data. Statistical Applications in Genetics and Molecular Biology,
vol. 9, iss. 1, article 9.
}
\keyword{manip}

