% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{createBiopaxnew}
\alias{createBiopaxnew}
\title{Create a new Biopax model containing all database information.}
\usage{
createBiopaxnew(intIDs, pwdatabases)
}
\arguments{
\item{intIDs}{output list of genintIDs function.}

\item{pwdatabases}{vector indicating with which pathway database the
pathways should be determined; possible choices are "biocarta", "kegg",
"nci", "reactome".}
}
\value{
biopax object generated from the specified pathway databases.
}
\description{
This function creates a new biopax model depending on which 
pathway databases are chosen for analysis.
}
\keyword{manip}

