% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{createIntIDs}
\alias{createIntIDs}
\title{Create internal IDs.}
\usage{
createIntIDs(data_omics, PWinfo)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{PWinfo}{pathway database information from chosen pathway databases as
from loadPWs.}
}
\value{
list of internal IDs from specified pathway databases.
}
\description{
Create new internal ids in biopax$df:
}
\keyword{manip}

