% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_downstream_analysis.R
\name{enrichPWs}
\alias{enrichPWs}
\title{Pathway enrichment - downstream analysis.}
\usage{
enrichPWs(data_omics, method = "BH", alpha = 0.05, ...)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{method}{correction method for multiple testing correction as specified
in p.adjust documentation; default is Benjamini & Hochberg correction.}

\item{alpha}{significance level for pathway enrichment;
default is alpha = 0.05.}

\item{...}{further input parameters for multiple comparison adjustment.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function does pathway enrichment for pathways determined via identifyPWs
function.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"),
TFtargetdatabase = c("chea"))
\donttest{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = "Genelists")
}
\dontrun{
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
}
}
\keyword{manip}

