% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_upstream_analysis.R
\name{enrichTFs}
\alias{enrichTFs}
\title{Transcription factor enrichment - upstream analysis.}
\usage{
enrichTFs(data_omics, method = "BH", alpha = 0.05, ...)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{method}{correction method for multiple testing correction as specified
in p.adjust documentation; default is Benjamini & Hochberg correction.}

\item{alpha}{significance level for transcription factor enrichment;
default is alpha = 0.05.}

\item{...}{further input parameters for multiple comparison adjustment.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function does transcription factor enrichment for transcription factor
identified to be upstream of the diff. expressed genes/transcripts in the
omics data set imported with readOmics function. In order to use this
function first read in the transcription factor target gene information via
readTFdata and identify the upstream TFs of the differentially expressed
genes/transcripts with the identifyTFs function.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"),
TFtargetdatabase = c("chea"))
\donttest{
setwd(system.file("extdata", package = "pwOmics"))
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = "Genelists")
}
\dontrun{
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichTFs(data_omics)
}
}
\keyword{manip}

