% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_integration_functions.R
\name{findSignalingAxes}
\alias{findSignalingAxes}
\title{Find downstream signaling axis.}
\usage{
findSignalingAxes(data_omics, phosphoprot, tpDS)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{phosphoprot}{character specifying the name of the phosphoprotein that is
selected as the starting point for downstream analysis.}

\item{tpDS}{integer specifying the time point considered for downstream analysis
of phosphoprotein data .}
}
\value{
list of downstream pathways identified for this time point of 
phosphoprotein measurement with sublists containing information about the 
transcription factor, the regulation, the target genes of these transcription 
factors and the matching transcripts.
}
\description{
This function determines the regulated downstream structures of a selected
phosphoprotein at a specified time point.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics, 
loadgenelists = system.file("extdata/Genelists", package = "pwOmics"))
setwd(system.file("extdata/Genelists", package = "pwOmics"))
\dontrun{
data_omics_plus = identifyPR(data_omics_plus)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyPWTFTGs(data_omics)
data_omics = identifyRsofTFs(data_omics, noTFs_inPW = 1, order_neighbors = 10)
SYK_axis = findSignalingAxes(data_omics,phosphoprot = "SYK", tpDS = 2)
}
}
\keyword{manip}
