% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_integration_functions.R
\name{generate_DSSignalingBase}
\alias{generate_DSSignalingBase}
\title{Generate a folder with downstream information about all phosphoproteins.}
\usage{
generate_DSSignalingBase(data_omics, timepoints = c(0.25, 1, 4, 8, 13, 18,
  24))
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{timepoints}{integer vector specifying the timepoints of interest for
downstream analysis.}
}
\value{
Folder structure in working directory, containing phosphoprotein
downstream information in individual folders.
}
\description{
This function generates a folder structure with an RData object and csv
tables for each timepoint sepcified for each phosphoprotein considered in
the data_omics object.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
\dontrun{
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics, 
loadgenelists = system.file("extdata/Genelists", package = "pwOmics"))
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyPWTFTGs(data_omics)
data_omics = identifyRsofTFs(data_omics, noTFs_inPW = 1, order_neighbors = 10)
setwd("~/Signaling_axes/")
generate_DSSignalingBase(data_omics, timepoints = c(0.25, 1, 4, 8, 13, 18, 24))
}
}
\keyword{manip}

