% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{genfullConsensusGraph}
\alias{genfullConsensusGraph}
\title{Combine SteinerNet with bipartite graph to get full consensus network.}
\usage{
genfullConsensusGraph(ST_net, ST_TFTG)
}
\arguments{
\item{ST_net}{steiner tree graph generated by SteinerTree_cons function.}

\item{ST_TFTG}{steiner tree graph extended with consensus target genes and 
the edges between TFs and target genes.}
}
\value{
igraph object of network comprising steiner tree graph and 
TF - target gene interactions.
}
\description{
Combine SteinerNet with bipartite graph to get full consensus network.
}
\keyword{manip}

