% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{genGenelists}
\alias{genGenelists}
\title{Generate genelists from pathway databases.}
\usage{
genGenelists(intIDs, pwdatabases)
}
\arguments{
\item{intIDs}{list containing Biopax model with newly generated internal IDs
 as processed with genintIDs function.
The components of the list are biopax models for "biocarta", "kegg", "nci",
 "reactome". In case a database was not chosen the list
entry contains a NA.}

\item{pwdatabases}{vector indicating with which pathway database the pathways
 should be determined; possible choices
are "biocarta", "kegg", "nci", "reactome".}
}
\value{
list of genelists of specified pathway databases.
}
\description{
This function generates genelists from the chosen pathway databases for
further processing in detPathways function.
}
\keyword{manip}

