% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{genIntIDs}
\alias{genIntIDs}
\title{Internal function for generation of pathway database specific internal IDs.}
\usage{
genIntIDs(data_omics, PWinfo, PWinfo_ind, PWDBname)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{PWinfo}{pathway database information from chosen pathway databases as
from loadPWs.}

\item{PWinfo_ind}{integer specifying element of loadPWs output matching the
chosen pathway database.}

\item{PWDBname}{character; pathway database name.}
}
\value{
data.table with newly generated internal IDs of biopax model.
}
\description{
Generates new internal ids (database-specific) in biopax$df.
}
\keyword{manip}
