% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_integration_functions.R
\name{get_matching_transcripts}
\alias{get_matching_transcripts}
\title{Summarize table of matching downstream transcripts.}
\usage{
get_matching_transcripts(data_omics, axis)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{axis}{list output of findSignalingAxes.}
}
\value{
dataframe containing the target genes of the axis matching the
transcript data in at least one of the time points in the first column, the
direction of their regulation inferred from the phosphoproteome data 
in the second column, the transcript regulation in the next columns (per time
point) and the summarized matched information in the following columns (per
time point).
}
\description{
This function returns the summarized table of matching transcripts information
based on the result list from the findSignalingAxes function.
}
\examples{
\dontrun{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics.newupdown")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics.newupdown"))
data_omics_plus = readPWdata(data_omics, 
loadgenelists = system.file("extdata/Genelists", package = "pwOmics.newupdown"))
}
\dontrun{
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics.newupdown"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyPWTFTGs(data_omics)
data_omics = identifyRsofTFs(data_omics, noTFs_inPW = 1, order_neighbors = 10)
SYK_axis = findSignalingAxes(data_omics,phosphoprot = "SYK", tpDS = 2)
SYK_transcripts = get_matching_transcripts(data_omics, SYK_axis)
}
}
\keyword{manip}

