% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{getAlias_Ensemble}
\alias{getAlias_Ensemble}
\title{Get Gene Symbols from Ensemble Gene Accession IDs.}
\usage{
getAlias_Ensemble(ids)
}
\arguments{
\item{ids}{vector of Ensemble Gene Accession IDs.}
}
\value{
ids character vector of gene symbols.
}
\description{
Get Gene Symbols from Ensemble Gene Accession IDs.
}
\keyword{manip}
