% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{getAliasfromSTRINGIDs}
\alias{getAliasfromSTRINGIDs}
\title{Map alias names to STRING IDs of consensus graph.}
\usage{
getAliasfromSTRINGIDs(data_omics, ST_net, updown = FALSE, phospho = TRUE,
  consSTRINGIDs, tps, string_db)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{ST_net}{steiner tree graph generated by SteinerTree_cons function.}

\item{updown}{boolean value; TRUE in case up- and downregulation should be
checked individually for intersection. Type of checking is defined with
parameter 'phospho'.}

\item{phospho}{boolean value; TRUE in case up- and downregulation should be
checked based on provided downstream phosphoprotein influence from 
identifyPR function; FALSE in case up- and downregulation should be checked
for without phosphoprotein database knowledge. Default is TRUE.}

\item{consSTRINGIDs}{first element of list generated by getConsensusSTRINGIDs 
function; a data.frame including the proteins to be considered as terminal 
nodes in Steiner tree with colnames ST_proteins and the corresponding STRING 
IDs in column 'STRING_id'.}

\item{tps}{integer specifying current timepoint under consideration.}

\item{string_db}{second element of list generated by getConsensusSTRINGIDs 
function; species table (for human) of STRING database.}
}
\value{
igraph object with alias name annotation.
}
\description{
Map alias names to STRING IDs of consensus graph.
}
\keyword{manip}
