% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getBiopaxModel}
\alias{getBiopaxModel}
\title{Get upstream regulators of identified transcription factors.}
\usage{
getBiopaxModel(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
biopax model generated as consensus biopax models from all 
pathway databases selected for analysis.
}
\description{
Get upstream regulators of identified transcription factors.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"), 
TFtargetdatabase = c("chea"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,  
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
getBiopaxModel(data_omics)
}
}
\keyword{manip}

