% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getDS_PWs}
\alias{getDS_PWs}
\title{Get downstream analysis pathways.}
\usage{
getDS_PWs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
list of length = number of protein time points, each element
containing a data frame with the pathway IDs in the generated biopax model,
corresponding pathway names and flags if those pathways are enriched
(1 = enriched, NA = not enriched).
}
\description{
This function returns pathways identified in the downstream analysis
containing the significantly abundant proteins.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
getDS_PWs(data_omics)
}
}
\keyword{manip}

