% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getDS_TFs}
\alias{getDS_TFs}
\title{Get downstream analysis transcription factors in pathways.}
\usage{
getDS_TFs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
list of length = number of protein time points, each element
containing a character vector with identified transcription factors.
}
\description{
This function returns the genes identified in the downstream analysis and
a column indicating if the genes are transcription factors.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
getDS_TFs(data_omics)
}
}
\keyword{manip}

