% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{getFCsplines}
\alias{getFCsplines}
\title{Get fold change splines.}
\usage{
getFCsplines(data_omics, nodes, nodetype)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{nodes}{character vector of nodes the fold change splines should be 
calculated for.}

\item{nodetype}{character indicating to calculate splines for "proteins" or 
"genes".}
}
\value{
splines values used in dynamic analysis.
}
\description{
Calculate the splines used for the dynamic analysis.
}
\keyword{manip}
