% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getOmicsallGeneIDs}
\alias{getOmicsallGeneIDs}
\title{Get all gene IDs.}
\usage{
getOmicsallGeneIDs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
all gene IDs.
}
\description{
This function returns the gene IDs of all genes (transcripts) measured.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"),
TFtargetdatabase = c("chea"))
getOmicsallGeneIDs(data_omics)
}
\keyword{manip}

