% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getOmicsallProteinIDs}
\alias{getOmicsallProteinIDs}
\title{Get all protein IDs.}
\usage{
getOmicsallProteinIDs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
all protein IDs.
}
\description{
This function returns the protein IDs of all proteins measured.
}
\examples{
\dontrun{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
getOmicsallProteinIDs(data_omics)
}
}
\keyword{manip}

