% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getProteinIntersection}
\alias{getProteinIntersection}
\title{Get protein intersection for the omics data on the different time points.}
\usage{
getProteinIntersection(data_omics, tp_prot, tp_genes)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{tp_prot}{numeric integer defining protein timepoint measurement chosen
for comparison.}

\item{tp_genes}{numeric integer defining gene/transcript timepoint
measurement chosen for comparison.}
}
\value{
list with three elements: 1) character vector of protein IDs
identified in both upstream and downstream analysis 2) protein time point
3) gene/transcript time point.
}
\description{
The timepoints or measurement names for comparison have to be defined in
tp_prot and tp_genes as given in the readOmics function.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE,
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
getProteinIntersection(data_omics, tp_prot = 4, tp_genes = 4)
}
}
\keyword{manip}

