% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{gettpIntersection}
\alias{gettpIntersection}
\title{Get omics data intersection on the three levels.}
\usage{
gettpIntersection(data_omics, only_enriched = FALSE)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{only_enriched}{Boolean value defining if transcription factors should
be identified only for enriched pathways (TRUE); or for all identified
pathways (FALSE); default is FALSE.}
}
\value{
list with three elements:
1) protein intersection
2) transcription factor intersection
3) gene intersection
each element contains a list with overlapping time points of both upstream
and downstream analyses.
}
\description{
Get intersection for the omics data on all three levels (proteins, TFs,
genes) on corresponding time points.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE,
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
gettpIntersection(data_omics)
}
}
\keyword{manip}

