% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getUS_PWs}
\alias{getUS_PWs}
\title{Get upstream pathways of identified transcription factors.}
\usage{
getUS_PWs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
list of length = number of gene/transcript time points, each element
containing a list of transcription factors; these transcription factor
elements contain data frame with pathway IDs and pathway names.
}
\description{
Get upstream pathways of identified transcription factors.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"), 
TFtargetdatabase = c("chea"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,  
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE, 
noTFs_inPW = 1, order_neighbors = 10)
getUS_PWs(data_omics)
}
}
\keyword{manip}

