% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_downstream_analysis.R
\name{identifyPR}
\alias{identifyPR}
\title{Identify phosphorylation regulation influence downstream}
\usage{
identifyPR(data_omics_plus)
}
\arguments{
\item{data_omics_plus}{output list of readPWdata function; first element
contains an OmicsData object, secons element the genelist data corresponding 
to the selected pathway database.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD, 
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function identifies the downstream regulation influence
of phosphoprotein regulation for further downstream analysis steps.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics"))
data_omics_plus = identifyPR(data_omics_plus)
}
\keyword{manip}
