% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_downstream_analysis.R
\name{identifyPWTFTGs}
\alias{identifyPWTFTGs}
\title{Identify TFs in enriched pathways and their target genes - downstream
analysis.}
\usage{
identifyPWTFTGs(data_omics, only_enriched = TRUE)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{only_enriched}{boolean value defining if transcription factors should
be identified only for enriched pathways (TRUE); or for all identified
pathways (FALSE); default is TRUE.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function identifies the transcription factors being part of the enriched
pathways of downstream analysis. Subsequently it finds the target genes of
these transcription factors from the selected TF-target gene database.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"),
TFtargetdatabase = c("chea"))
\donttest{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
}
\dontrun{
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
identifyPWTFTGs(data_omics, only_enriched = FALSE)
}
}
\keyword{manip}

