% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_upstream_analysis.R
\name{identifyRsofTFs}
\alias{identifyRsofTFs}
\title{Identify regulators of transcription factors - upstream analysis.}
\usage{
identifyRsofTFs(data_omics, noTFs_inPW = 2, order_neighbors = 6)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{noTFs_inPW}{integer; only regulators in upstream pathways with more
than this number of TFs are identified.}

\item{order_neighbors}{integer specifiying the order of the neighborhood:
order 1 is TF plus its immediate neighbors.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function identifies the regulators upstream of the identified 
transcription factors in upstream analysis.
Converting the pathway information to a regulatory graph needs some time...
Warnings regarding the skipping of edges in building the regulatory graph can 
be ignored.
}
\examples{
\dontrun{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics.newupdown")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics.newupdown"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics.newupdown")) 
}
\dontrun{
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics.newupdown"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyRsofTFs(data_omics, 
noTFs_inPW = 1, order_neighbors = 10)
}
}
\keyword{manip}

