% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_upstream_analysis.R
\name{identifyTFs}
\alias{identifyTFs}
\title{Transcription factor identification.}
\usage{
identifyTFs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
 Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function identifies the upstream transcription factors of the provided
 gene IDs.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"), 
TFtargetdatabase = c("chea"))
\donttest{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics, 
loadgenelists = "Genelists")
}
\dontrun{
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
}
}
\keyword{manip}

