% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{identPWsofTFs}
\alias{identPWsofTFs}
\title{Identification of pathways containing the transcription factors identified in 
upstream analysis}
\usage{
identPWsofTFs(genelists, tps_TFs)
}
\arguments{
\item{genelists}{data.table as read/loaded by loadGenelist function.}

\item{tps_TFs}{data.table of upstream transcription factors 
as returned from identTFs function.}
}
\value{
list with first element being a pathway list and second being a 
pathway dataframe of pathways including the TFs of the specified timepoint.
}
\description{
Identification of pathways containing the transcription factors identified in 
upstream analysis
}
\keyword{manip}

