% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{identTFs}
\alias{identTFs}
\title{This function provides a data.table of upstream transcription factors.}
\usage{
identTFs(data_omics, glen)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{glen}{numeric value; identifier for current timepoint.}
}
\value{
data.table of upstream TFs.
}
\description{
This function provides a data.table of upstream transcription factors.
}
\keyword{manip}

