% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{identTFTGsinPWs}
\alias{identTFTGsinPWs}
\title{Prepare OmicsData object for pathway information.}
\usage{
identTFTGsinPWs(data_omics, temp_genelist)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{temp_genelist}{dataframe of unique gene IDs in enriched/not enriched
PWs.}
}
\value{
list with first element being a genelist of the pathways and second
being a target gene list of TFs.
}
\description{
This function identifies the TFs in the pathway genes and determines their
target genes on basis of the given (chosen) TF-target database(s).
}
\keyword{manip}

