% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{infoConsensusGraph}
\alias{infoConsensusGraph}
\title{Add Consensus Graph information.}
\usage{
infoConsensusGraph(data_omics, consensusGraph, phosphotab)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{consensusGraph}{result from static analysis: consensus graph generated 
by staticConsensusNet function.}

\item{phosphotab}{dataframe with phosphoprotein information annotated in 
columns 'Gene.names', 'Amino acid', 'Position' (of phosphosite).}
}
\value{
graph of igraph class containing complemented consensus graph
information
}
\description{
Adds phosphoprotein information based on phosphoprotein data table and 
redraws Consensus Graph edges
}
\examples{
\donttest{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics.newupdown")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics.newupdown"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics.newupdown")) 
}
\dontrun{
data_omics_plus = identifyPR(data_omics_plus) 
setwd(system.file("extdata/Genelists", package = "pwOmics.newupdown"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyRsofTFs(data_omics, 
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics)
statConsNet = staticConsensusNet(data_omics)
}
}
\keyword{manip}

