% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{loadPWs}
\alias{loadPWs}
\title{Load pathway database information.}
\usage{
loadPWs(pwdatabases, biopax_level)
}
\arguments{
\item{pwdatabases}{vector indicating with which pathway database the 
pathways should be determined; possible choices
are "biocarta", "kegg", "nci", "reactome".}

\item{biopax_level}{integer indicating biopax level of pathway database 
information to be retrieved.}
}
\value{
list of biopax model corresponding to specified pathway databases.
}
\description{
This function loads the pathway information from pathway databases. 
Needed in the identifyPWs function.
}
\keyword{manip}
