% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_visualization_functions.R
\name{plotConsDynNet}
\alias{plotConsDynNet}
\title{Plot inferred net based on analysis analysis.}
\usage{
plotConsDynNet(dynConsensusNet, sig.level, clarify = "TRUE",
  layout = layout.fruchterman.reingold, ...)
}
\arguments{
\item{dynConsensusNet}{result of dynamic analysis: inferred net generated by
consDynamicNet function.}

\item{sig.level}{significance level for plotting edges in network
(between 0 and 1).}

\item{clarify}{indicating if unconnected nodes should be removed;
default = "TRUE".}

\item{layout}{igraph layout parameter; default is
layout.fruchterman.reingold.}

\item{...}{further plotting/legend parameters.}
}
\value{
pdf file in current working directory.
}
\description{
Dynamic analysis result is plotted to pdf file stored in current working
directory.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE,
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
statConsNet = staticConsensusNet(data_omics)
dynConsNet = consDynamicNet(data_omics, statConsNet)
plotConsDynNet(dynConsNet, sig.level = 0.8)
}
}
\keyword{manip}

