% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_vizualisation_functions.R
\name{plotConsensusGraph}
\alias{plotConsensusGraph}
\title{Plot consensus graph(s) from static analysis.}
\usage{
plotConsensusGraph(consensusGraphs, data_omics, ...)
}
\arguments{
\item{consensusGraphs}{result from static analysis: consensus graph generated 
by staticConsensusNet function.}

\item{data_omics}{OmicsData object.}

\item{...}{further plotting/legend parameters.}
}
\value{
pdf file in current working directory.
}
\description{
Consensus graph(s) plotted to pdf file stored in current working directory.
}
\examples{
\donttest{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics.newupdown")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics.newupdown"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics.newupdown")) 
}
\dontrun{
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics.newupdown"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyRsofTFs(data_omics, 
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics)
statConsNet = staticConsensusNet(data_omics)
plot(ConsensusGraph, data_omics)
}
}
\keyword{manip}

