% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_vizualisation_functions.R
\name{plotConsensusProfiles}
\alias{plotConsensusProfiles}
\title{Plot consensus graph profiles of static consensus molecules.}
\usage{
plotConsensusProfiles(consensusGraphs, data_omics, subsel = TRUE, ...)
}
\arguments{
\item{consensusGraphs}{result from static analysis: consensus graph generated
by staticConsensusNet function.}

\item{data_omics}{OmicsData object.}

\item{subsel}{character vector of selected consensus molecules for plotting;
if TRUE all consensus molecules are plotted}

\item{...}{further plotting/legend parameters.}
}
\value{
pdf file in current working directory.
}
\description{
Consensus graph profiles of static consensus molecules
plotted as heatmap to pdf file stored in current working directory.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = FALSE)
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE,
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
statConsNet = staticConsensusNet(data_omics)
plotConsensusProfiles(statConsNet)
}
}
\keyword{manip}

