% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_vizualisation_functions.R
\name{plotTimeProfileClusters}
\alias{plotTimeProfileClusters}
\title{Plot time profile clusters of dynamic analysis result.}
\usage{
plotTimeProfileClusters(fuzzed_matsplines)
}
\arguments{
\item{fuzzed_matsplines}{result of dynamic analysis: inferred net generated
by dynamicConsensusNet function.}

\item{...}{further plotting/legend parameters.}
}
\value{
pdf file in current working directory.
}
\description{
Plot time profile clusters of dynamic analysis result.
}
\examples{
#please run with whole database files (prepared according to vignette)
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"),
TFtargetdatabase = c("chea", "pazar"))
\dontrun{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,
loadgenelists = "Genelists")
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
data_omics = identifyRsofTFs(data_omics, only_enriched = FALSE,
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics, only_enriched = FALSE)
statConsNet = staticConsensusNet(data_omics)
dynConsNet = dynamicConsensusNet(data_omics, statConsNet)
timeprof = clusterTimeProfiles(dynConsNet)
plotTimeProfileClusters(timeprof)
}
}
\keyword{manip}

