% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{predictFCvals}
\alias{predictFCvals}
\title{Prediction of continous data points via smoothing splines.}
\usage{
predictFCvals(data_omics, nopredpoints, splineslist, title)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{nopredpoints}{integer number; how many timpoints should be predicted?}

\item{splineslist}{list of protein or gene nodes for which splines were 
generated: output of getFCsplines function.}

\item{title}{character vector specifying name of title.}
}
\value{
list of splines matrix values and calculated times.
}
\description{
Prediction of continous data points via smoothing splines.
}
\keyword{manip}

