% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{preparePWinfo}
\alias{preparePWinfo}
\title{Prepare OmicsData object for pathway information.}
\usage{
preparePWinfo(data_omics, plen)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{plen}{integer indicating the timepoint currently investigated.}
}
\value{
list of OmicsData object, current timepoint and pathways of interest.
}
\description{
This function prepares the OmicsData object for the identified pathway
information.
}
\keyword{manip}

