% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_print.R
\name{print.OmicsData}
\alias{print.OmicsData}
\title{Print an OmicsData object.}
\usage{
\method{print}{OmicsData}(x, ...)
}
\arguments{
\item{x}{an OmicsData object to print.}

\item{...}{further arguments to be passed to print.}
}
\value{
prints OmicsData object.
}
\description{
Print an OmicsData object.
}
\keyword{manip}

