% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{PWidentallprots}
\alias{PWidentallprots}
\title{Identification of pathwayIDs and pathway names for all proteins.}
\usage{
PWidentallprots(data_omics, genelists)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{genelists}{lists of genelists from chosen pathway databases.}
}
\value{
OmicsData object with identified pathway IDs for list of all
proteins.
}
\description{
Identification of pathwayIDs and pathway names for all proteins.
}
\keyword{manip}

