% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{PWidentallprotssub}
\alias{PWidentallprotssub}
\title{Internal subfunction for all protein pathway identification.}
\usage{
PWidentallprotssub(data_omics, genelists, genelist_ind, datab)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{genelists}{lists of genelists from chosen pathway databases}

\item{genelist_ind}{integer specifying pathway database genelist matching;
1 = biocarta, 2 = kegg, 3 = nci, 4 = reactome.}

\item{datab}{character vector indicating database name for message.}
}
\value{
OmicsData object with identified pathways for each protein.
}
\description{
Internal subfunction for all protein pathway identification.
}
\keyword{manip}

