% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{PWidenttps}
\alias{PWidenttps}
\title{Identification of pathwayIDs and pathway names for proteins at individual
timepoints.}
\usage{
PWidenttps(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
data_omics OmicsData object with all pathways identified for the
individual timepoints.
}
\description{
Take the identified pathways from the list of all proteins and transfer this
information for the individual timepoints.
}
\keyword{manip}

