\name{pwIntOmics-package}
\alias{pwIntOmics-package}
\alias{pwIntOmics}
\docType{package}
\title{Pathway-based data integration of omics data}
\description{pwIntOmics performs pathway-based level-specific data comparison 
    of matching omics data sets based on pre-analysed user-specified lists of
    differential genes/transcripts and proteins. A separate downstream analysis
    of proteomic data including pathway identification and enrichment analysis, 
    transcription factor identification and target gene identification is 
    opposed to the upstream analysis starting with gene or transcript 
    information as basis for identification of upstream transcription factors 
    and regulators. The cross-platform comparative analysis allows for 
    comprehensive analysis of single time point experiments and time-series 
    experiments by providing static and dynamic analysis tools for data 
    integration.}
\details{
\tabular{ll}{
Package: \tab pwIntOmics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-29\cr
License: \tab GLP-3\cr
}
}
\author{Astrid Wachter
Maintainer: Astrid Wachter <Astrid.Wachter@med.uni-goettingen.de>
}
\keyword{ package }

