% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_read_data.R
\name{readOmics}
\alias{readOmics}
\title{Read in omics data.}
\usage{
readOmics(tp_prots, tp_genes, omics, PWdatabase, TFtargetdatabase)
}
\arguments{
\item{tp_prots}{numeric vector of protein timepoints used in experiment; 
in case of single time point experiments simply assign an experiment number
(e.g. 1).}

\item{tp_genes}{numeric vector of gene/transcript timepoints used in 
experiment; in case of single time point experiments simply assign an 
experiment number (e.g. 1).}

\item{omics}{list containing protein and gene IDs and fold changes: 
Input are two lists, one for protein data, one for gene data,
both including a vector of all measured IDs as first element and a list as 
second element including for each tp a dataframe with IDs and foldchanges 
per timepoint.}

\item{PWdatabase}{character vector of pathway database names which should be
used for pathway identification, possible choices are "biocarta", "kegg", 
"nci", "reactome".}

\item{TFtargetdatabase}{character vector of TF target database names which 
should be used for transcription factor/target gene identification,
possible choices are "chea", "pazar", "userspec". In case the user is able 
to provide an own list of transcription factor target gene matches, he can
indicate this via "userspec".}
}
\value{
OmicsData object:
list of 4 elements (OmicsD, PathwayD, TFtargetsD, 
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function reads omics data: differentially expressed genes and relatively
differentially abundant proteins with corresponding fold changes for each 
time point.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"), 
TFtargetdatabase = c("chea"))

\dontrun{
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("chea", "pazar", "userspec"))
}
}
\keyword{manip}

