% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_read_data.R
\name{readPhosphodata}
\alias{readPhosphodata}
\title{Reads in phosphoprotein downstream regulation information.}
\usage{
readPhosphodata(data_omics, phosphoreg)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{phosphoreg}{txt file with 2 columns, first providing gene names, second
giving -1/1 annotation whether downstream regulation of phosphoprotein
is inhibiting or activating, respectively.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD, 
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data;
PhosphoD containing optionally phosphoprotein downstream regulation annotation.
}
\description{
This function reads in phosphoprotein downstream regulation
information from a txt file.
}
\examples{
\dontrun{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
}
}
\keyword{manip}
