% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_read_data.R
\name{readTFdata}
\alias{readTFdata}
\title{Reads in chosen transcription factor target database information.}
\usage{
readTFdata(data_omics, TF_target_path, TF_filter_threshold = 0)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{TF_target_path}{character vector indicating path of the txt file of
matching transcription factors and target genes; the file should be a txt
file with first column transcription factors and second column target gene
symbols without a header.}

\item{TF_filter_threshold}{integer defining a threshold number to
filter out those transcription factors having higher numbers of target genes
than 'TF_filter_threshold' from the further analysis}
}
\value{
OmicsData object - a list containing information about the user data
(timepoints, IDs, fold changes) and the selected databases chosen for the
analysis.

OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function reads in transcription factor information given the selected
transcription factor target gene database. The information is downloaded
via the AnnotationHub package and merged, if necessary.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"),
TFtargetdatabase = c("chea"))
data_omics = readTFdata(data_omics)
data_omics[[3]]
}
\keyword{manip}

