% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{readTFtargets}
\alias{readTFtargets}
\title{Read in matching transcription factor target gene pairs.}
\usage{
readTFtargets(data_omics, TF_target_path)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{TF_target_path}{path of txt file containing the transcription factor
target gene information as specified above.}
}
\value{
data frame with user-specified TF target gene information.
}
\description{
In case the user is able to provide a file with transcription factor - 
target gene matches (e.g. from TRANSFAC database) this function can read in 
the information.
The file needs to be a txt file with first column transcription factors and 
second column target gene symbols without a header.
}
\keyword{manip}
