% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{selectPWsofTFs}
\alias{selectPWsofTFs}
\title{Select pathways with more than x TFs}
\usage{
selectPWsofTFs(pathway_list, pathway_frame, noTFs_inPW)
}
\arguments{
\item{pathway_list}{first element of list returned from identPWsofTFs
function; contains a list of pathways.}

\item{pathway_frame}{second element of list returned from identPWsofTFs
function; contains a data.frame of pathways.}

\item{noTFs_inPW}{numeric value specifying number of TFs being at least
part of the pathway.}
}
\value{
list of pathways with more than x TFs.
}
\description{
Select pathways with more than x TFs
}
\keyword{manip}

