% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{SteinerTree_cons}
\alias{SteinerTree_cons}
\title{Steiner tree algorithm.}
\usage{
SteinerTree_cons(terminal_nodes, PPI_graph, run_times)
}
\arguments{
\item{terminal_nodes}{character vector of final nodes used for generation
of Steiner tree.}

\item{PPI_graph}{igraph object; graph should be connected and have
undirected edges.}

\item{run_times}{integer specifying number of times to run SP Steiner tree
algorithm to find minimal graph.}
}
\value{
igraph object including Steiner tree.
}
\description{
Use this function to get the Steiner tree based on the STRING protein-protein 
interaction database.
}
\keyword{manip}
