% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_vizualisation_functions.R
\name{temp_correlations}
\alias{temp_correlations}
\title{Plot temporal correlations of phosphoprotein expression levels and affected 
downstream transcripts.}
\usage{
temp_correlations(ConsensusGraph, timepointsprot, timepointstrans,
  foldername = "ProtCons_", trans_sign = "signif_single.csv",
  trans_sign_names, phospho_sign = "mat_phospho.csv", phospho_sign_names)
}
\arguments{
\item{ConsensusGraph}{result from static analysis: consensus graph generated 
by staticConsensusNet function.}

\item{timepointsprot}{numeric vector with measurement time points in 
phosphoproteome data set, which should be used for correlation plots}

\item{timepointstrans}{numeric vector with measurement time points in
transcriptome data set, which should be used for correlation plots}

\item{foldername}{character vector specifying the name of the folder that
will be generated for the temporal correlations}

\item{trans_sign}{character vector specifying a tab-delimited file with the
transcriptome expression levels for all time points in timepointstrans}

\item{trans_sign_names}{character vector specifying column names in the
transcriptome file corresponding to the expression levels at different time
points.}

\item{phospho_sign}{character vector specifying a tab-delimited file with the
phosphoproteome information (columns 'Gene.names', 'Amino.acid', 'Position',
'Multiplicity') and expression levels for all time points in timepointstrans}

\item{phospho_sign_names}{character vector specifying column names in the
phosphoproteome file corresponding to the expression levels at different time
points.}

\item{...}{further plotting/legend parameters.}
}
\value{
pdf file in current working directory.
}
\description{
This function plots an overview of consensus phosphoprotein expression level 
correlations with those transcripts that are affected downstream in a newly
generated folder in the working directory. The following 
additional information needs to be provided: Phosphorylation information amino
acid, position and multiplicity with the expression levels for each
of the time points the correlations should be plotted for. Furthermore 
data tables for fold changes/ratios of phosphoproteins and transcripts that 
are part of the data_omics object.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
\dontrun{
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics")) 
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyRsofTFs(data_omics, 
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics)
statConsNet = staticConsensusNet(data_omics)
temp_correlations(statConsNet, timepointsprot = c(1,4,8,13,18,24),
timepointstrans = c(1,4,8,13,18,24), foldername = "ProtCons_", 
trans_sign = system.file("extdata", "signif_single.csv", package = "pwOmics") 
trans_sign_names = c("FC_1",  "FC_2", "FC_3", "FC_4"), 
phospho_sign = system.file("extdata", "mat_phospho.csv", package = "pwOmics") 
phospho_sign_names = c("Rat1", "Rat2", "Rat3", "Rat4")) )
}

}
\keyword{manip}

