% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{TFidentallgenes}
\alias{TFidentallgenes}
\title{Identification of upstream transcription factors for all genes.}
\usage{
TFidentallgenes(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
data_omics OmicsData object with upstream TFs identified for all
genes.
}
\description{
Identification of upstream transcription factors for all genes.
}
\keyword{manip}

