% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{TFidenttps}
\alias{TFidenttps}
\title{Identification of upstream transcription factors.}
\usage{
TFidenttps(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
data_omics OmicsData object with upstream TFs of differentially 
expressed genes for separate timepoints identified.
}
\description{
Identification of upstream transcription factors for the differentially 
expressed genes of the different timepoints.
}
\keyword{manip}
