% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{getbipartitegraphInfo}
\alias{getbipartitegraphInfo}
\title{Get TF-target gene information for the consensus graph.}
\usage{
getbipartitegraphInfo(data_omics, tps)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{tps}{integer specifying current timepoint under consideration.}
}
\value{
list of transcription factor target gene interactions.
}
\description{
Get TF-target gene information for the consensus graph.
}
\keyword{manip}

