% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_consensus_analysis.R
\name{getConsensusSTRINGIDs}
\alias{getConsensusSTRINGIDs}
\title{Get consensus graph in STRING IDs.}
\usage{
getConsensusSTRINGIDs(data_omics, tps, string_db)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{tps}{integer specifying current timepoint under consideration.}

\item{string_db}{STRING_db object.}
}
\value{
igraph object consensus graph with STRING IDs (only including 
proteins and transcription factors).
}
\description{
Get consensus graph in STRING IDs.
}
\keyword{manip}

