% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getDS_PWs}
\alias{getDS_PWs}
\title{Get downstream analysis pathways.}
\usage{
getDS_PWs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
list of length = number of protein time points, each element
containing a data frame with the pathway IDs in the generated biopax model and
corresponding pathway names.
}
\description{
This function returns pathways identified in the downstream analysis 
containing the significantly abundant proteins.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
\dontrun{
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics, 
loadgenelists = system.file("extdata/Genelists", package = "pwOmics"))
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyPWTFTGs(data_omics)
getDS_PWs(data_omics)
}
}
\keyword{manip}

