% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_get_functions.R
\name{getDS_TFs}
\alias{getDS_TFs}
\title{Get downstream analysis transcription factors in pathways.}
\usage{
getDS_TFs(data_omics)
}
\arguments{
\item{data_omics}{OmicsData object.}
}
\value{
list of length = number of protein time points, each element
containing a character vector with identified transcription factors.
}
\description{
This function returns the genes identified in the downstream analysis and 
a column indicating if the genes are transcription factors.
}
\examples{
\donttest{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics.newupdown")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics.newupdown"))
data_omics_plus = readPWdata(data_omics, 
loadgenelists = system.file("extdata/Genelists", package = "pwOmics.newupdown"))
}
\dontrun{
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics.newupdown"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyPWTFTGs(data_omics)
getDS_TFs(data_omics)
}
}
\keyword{manip}

