% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_downstream_analysis.R
\name{identifyPWs}
\alias{identifyPWs}
\title{Identify pathway IDs and pathway names of differentially abundant proteins}
\usage{
identifyPWs(data_omics_plus)
}
\arguments{
\item{data_omics_plus}{output list of readPWdata function; first element
contains an OmicsData object, secons element the genelist data corresponding 
to the selected pathway database.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD, 
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function identifies the pathways of the differentially abundant 
phosphoproteins dependent on the chosen database. 
Requires rBiopaxParser package. Takes a 
lot of time for a high number of proteins and/or if all databases are chosen.
First, chosen databases are loaded, then new internal pathway IDs are 
generated. 
Afterwards the genelists of the different databases are loaded or generated, 
depending on the loadgenelists option. After pathway identification for the 
reference time point, also pathway identification for different time points 
is performed. Pathway ID mapping takes some time, especially for such big 
databases as reactome, so use savegenelists and loadgenelists for easier and
faster usage...
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
\dontrun{
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics"))
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWs(data_omics_plus)
}
}
\keyword{manip}

