% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_upstream_analysis.R
\name{identifyRsofTFs}
\alias{identifyRsofTFs}
\title{Identify regulators of enriched transcription factors - upstream analysis.}
\usage{
identifyRsofTFs(data_omics, only_enriched = TRUE, noTFs_inPW = 2,
  order_neighbors = 6)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{only_enriched}{boolean value defining if transcription factors should
be identifies only for enriched pathways (TRUE); or for all identified
pathways (FALSE); default is TRUE.}

\item{noTFs_inPW}{integer; only regulators in upstream pathways with more
than this number of TFs are identified.}

\item{order_neighbors}{integer specifiying the order of the neighborhood:
order 1 is TF plus its immediate neighbors.}
}
\value{
OmicsData object: list of 4 elements (OmicsD, PathwayD, TFtargetsD,
Status); OmicsD containing omics data set + results (after analysis);
PathwayD containing selected pathway databases + biopax model;
TFtargetsD containing selected TF target gene databases + TF target gene data.
}
\description{
This function identifies the regulators upstream of the enriched/identified 
transcription factors in upstream analysis.
Converting the pathway information to a regulatory graph needs some time...
Warnings regarding the skipping of edges in building the regulatory graph can 
be ignored.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"), 
TFtargetdatabase = c("chea"))
\donttest{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics,  
loadgenelists = FALSE)
}
\dontrun{
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = enrichPWs(data_omics)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
identifyRsofTFs(data_omics, only_enriched = FALSE, noTFs_inPW = 1, 
order_neighbors = 10)
}
}
\keyword{manip}

