% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{identTFs}
\alias{identTFs}
\title{This function provides a data.table of upstream transcription factors and the
flag for enrichment.}
\usage{
identTFs(data_omics, glen)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{glen}{numeric value; identifier for current timepoint.}
}
\value{
data.table of upstream TFs and an enrichment flag.
}
\description{
This function provides a data.table of upstream transcription factors and the
flag for enrichment.
}
\keyword{manip}

