% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_helper_functions.R
\name{loadGenelists}
\alias{loadGenelists}
\title{Loading of genelists}
\usage{
loadGenelists()
}
\value{
genelist of specified pathway database.
}
\description{
This function automatically loads the genelists corresponding to the selected
pathway databases stored as RData file in the current working directory.
}
\keyword{manip}

