% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwOmics_vizualisation_functions.R
\name{plotConsDynNet}
\alias{plotConsDynNet}
\title{Plot inferred net based on analysis analysis.}
\usage{
plotConsDynNet(dynConsensusNet, sig.level, clarify = "TRUE",
  layout = layout.fruchterman.reingold, ...)
}
\arguments{
\item{dynConsensusNet}{result of dynamic analysis: inferred net generated by 
consDynamicNet function.}

\item{sig.level}{significance level for plotting edges in network 
(between 0 and 1).}

\item{clarify}{indicating if unconnected nodes should be removed; 
default = "TRUE".}

\item{layout}{igraph layout parameter; default is 
layout.fruchterman.reingold.}

\item{...}{further plotting/legend parameters.}
}
\value{
pdf file in current working directory.
}
\description{
Dynamic analysis result is plotted to pdf file stored in current working 
directory.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24), 
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta", "kegg", "nci", "reactome"), 
TFtargetdatabase = c("userspec"))
data_omics = readPhosphodata(data_omics, 
phosphoreg = system.file("extdata", "phospho_reg_table.txt", 
package = "pwOmics")) 
data_omics = readTFdata(data_omics, 
TF_target_path = system.file("extdata", "TF_targets.txt", 
package = "pwOmics"))
data_omics_plus = readPWdata(data_omics,  
loadgenelists = system.file("extdata/Genelists", package = "pwOmics")) 
\dontrun{
data_omics_plus = identifyPR(data_omics_plus)
setwd(system.file("extdata/Genelists", package = "pwOmics"))
data_omics = identifyPWs(data_omics_plus)
data_omics = identifyTFs(data_omics)
data_omics = identifyRsofTFs(data_omics, 
noTFs_inPW = 1, order_neighbors = 10)
data_omics = identifyPWTFTGs(data_omics)
statConsNet = staticConsensusNet(data_omics)
dynConsNet = consDynamicNet(data_omics, statConsNet)
plotConsDynNet(dynConsNet, sig.level = 0.8)
}
}
\keyword{manip}
