% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pwOmics_read_data.R
\name{readPWdata}
\alias{readPWdata}
\title{Read in pathway database data needed for pathway identification.}
\usage{
readPWdata(data_omics, loadgenelists, biopax_level = 2)
}
\arguments{
\item{data_omics}{OmicsData object.}

\item{loadgenelists}{path of genelist RData files stored previously; all
genelists stored in this path are read in and used automatically if path is
given; if loadgenelists = FALSE, then genelists from pathway databases have
to be generated first.}

\item{biopax_level}{integer indicating biopax level of pathway database
information. default level is 2.}
}
\value{
list of OmicsData object and genelists for selected pathway databases.
}
\description{
This function reads pathway data of the chosen database(s) via the
AnnotationHub [1] package and rBiopaxParser [2] package.
Takes a lot of time for a high number of proteins and/or if all databases
are chosen.
First, chosen databases are retrieved, then new internal pathway IDs are
generated. Afterwards the genelists of the different databases are loaded or
generated, depending on the loadgenelists option. Pathway ID mapping takes
some time, especially for such big databases as reactome, so the genelists
are automatically stored in the current working folder and can be used
via loadgenelists in case you use this function again for easier and faster
usage...
Biopax level of retrieved databases is 2 by default.
}
\examples{
data(OmicsExampleData)
data_omics = readOmics(tp_prots = c(0.25, 1, 4, 8, 13, 18, 24),
tp_genes = c(1, 4, 8, 13, 18, 24), OmicsExampleData,
PWdatabase = c("biocarta"),
TFtargetdatabase = c("chea"))
\donttest{
data_omics = readTFdata(data_omics)
data_omics_plus = readPWdata(data_omics, loadgenelists = "Genelists")
data_omics_plus[[2]][[1]]
}
}
\references{
1. Morgan M, Carlson M, Tenenbaum D and Arora S. AnnotationHub:
Client to access AnnotationHub resources. R package version 1.99.75.

2. Kramer F, Bayerlova M, Klemm F, Bleckmann A and Beissbarth T.
rBiopaxParser - an R package to parse, modify and visualize BioPAX data.
}
\keyword{manip}

